refactor(sbase): Phase 5 — remove all legacy tracking tables and entities
Drop tracker_session, location_data, wifi_data, cell_location,
location_wifi_count tables via MIGRATION_20_21. Strip @Entity from
TrackerSession (kept as POJO, used by 41+ files) and DatabaseLocation.

- Rewrite TripDao: remove 9 UNION ALL queries, pure session_segment
- Refactor SessionTrackerComponent: remove SessionDataDao, write only
  SessionSegment; insert on init, update on disable
- Remove legacyDualWriteEnabled from settings layer
- Delete legacy DAOs: SessionDataDao, LocationDataDao, WifiDataDao,
  CellLocationDao, LocationWifiCountDao
- Delete legacy entities: DatabaseWifiData, DatabaseCellLocation,
  DatabaseLocationWifiCount
- Update 30+ files across statistics, impexp, activity, game, tracker,
  dashboard, app, testing-common modules
- Add Hilt bindings for CellSampleDao, WifiObservationDao,
  PressureSampleDao, StepIntervalDao, ActivitySnapshotDao,
  PersistenceErrorCollector
- Fix CellSampleDaoTest cellId Int→Long type mismatch
- Fix SessionTrackerComponentTest to expect update() not insert()

Build: assembleDebug ✓  Tests: tracker + sbase ✓

+++ sbase/src/test/java/com/adsamcik/tracker/sbase/CellSampleDaoTest.kt
@@ -37,7 +37,7 @@ class CellSampleDaoTest {
 
 	private fun createSample(
 		timeMs: Long = 1000L,
-		cellId: Int = 12345,
+		cellId: Long = 12345L,
 		lac: Int = 100,
 		mcc: Int = 230,
 		mnc: Int = 1,
@@ -157,7 +157,7 @@ class CellSampleDaoTest {
 	@Test
 	fun `getSamplesWithoutCoordinates respects limit`()  { runTest {
 		repeat(5) { i ->
-			dao.insert(createSample(timeMs = (i * 1000 + 1000).toLong(), cellId = i))
+			dao.insert(createSample(timeMs = (i * 1000 + 1000).toLong(), cellId = i.toLong()))
 		}
 
 		dao.getSamplesWithoutCoordinates(3) shouldHaveSize 3
